Get rid of Documentation/feature-removal.txt
This file has turned out to be a pure write-only file that causes merge
conflicts and has no actual redeeming features.

There is never any reason to add stuff to this idiotic file.  Either
something isn't getting used, and you should just remove it, or there is
no excuse for removing it in the first place.  Just stop the idiocy.

It has also been the excuse for just plain bad behavior ("Hey, I don't
like xyz, so let's mark it for removal" followed by "Hey, look, it been
in feature-removal.txt for six months now, so we should remove it").

The recent bogus setitimer() ABI change request was just the most recent
example of pointless and incorrect mental masturbation involving this
file, and I'm tired of the silly and pointless conflicts in the file.

This removal was discussed during the recent kernel summit, and while
Steven Rostedt suggested we should just enter the file itself in the
feature-removal schedule (to see if anybody who edits the file actually
looks at it), that's cute but non-productive.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>